doc:插件相关的常量位置修改并修改引用
