+++ app/renderer/src/main/src/pages/plugins/user/PluginUser.tsx
@@ -22,9 +22,8 @@ import {
     OutlineTrashIcon
 } from "@/assets/icon/outline"
 import {useMemoizedFn, useDebounceFn, useControllableValue, useInViewport, useLatest} from "ahooks"
-import {OnlineJudgment} from "../onlineJudgment/OnlineJudgment"
 import cloneDeep from "lodash/cloneDeep"
-import {PluginsContainer, PluginsLayout, defaultSearch, statusTag} from "../baseTemplate"
+import {PluginsContainer, PluginsLayout, statusTag} from "../baseTemplate"
 import {PluginFilterParams, PluginSearchParams, PluginListPageMeta} from "../baseTemplateType"
 import {SolidPluscircleIcon} from "@/assets/icon/solid"
 import {yakitNotify} from "@/utils/notification"
@@ -63,7 +62,7 @@ import {YakitHint} from "@/components/yakitUI/YakitHint/YakitHint"
 import {YakitCheckbox} from "@/components/yakitUI/YakitCheckbox/YakitCheckbox"
 import {PluginRecycleList} from "./PluginRecycleList"
 import {getRemoteValue, setRemoteValue} from "@/utils/kv"
-import {DefaultPublicStatusList, PluginGV} from "../builtInData"
+import {DefaultPublicStatusList, PluginGV, defaultSearch} from "../builtInData"
 import emiter from "@/utils/eventBus/eventBus"
 import {YakitRoute} from "@/routes/newRoute"
 import {showYakitModal} from "@/components/yakitUI/YakitModal/YakitModalConfirm"
