+++ app/renderer/src/main/src/pages/securityTool/newPortScan/NewPortScan.tsx
@@ -35,10 +35,9 @@ import {isEnpriTrace} from "@/utils/envfile"
 import {PluginExecuteResult} from "@/pages/plugins/operator/pluginExecuteResult/PluginExecuteResult"
 import {randomString} from "@/utils/randomUtil"
 import useHoldGRPCStream from "@/hook/useHoldGRPCStream/useHoldGRPCStream"
-import {defaultSearch} from "@/pages/plugins/baseTemplate"
 import {PluginLocalListDetails} from "@/pages/plugins/operator/PluginLocalListDetails/PluginLocalListDetails"
 import {PluginFilterParams, PluginSearchParams} from "@/pages/plugins/baseTemplateType"
-import {pluginTypeToName} from "@/pages/plugins/builtInData"
+import {defaultSearch, pluginTypeToName} from "@/pages/plugins/builtInData"
 import {defaultLinkPluginConfig} from "@/pages/plugins/utils"
 import {getLinkPluginConfig} from "@/pages/plugins/singlePluginExecution/SinglePluginExecution"
 import {RecordPortScanRequest, apiCancelPortScan, apiCancelSimpleDetect, apiPortScan, apiSimpleDetect} from "./utils"
